moved amp confighelper
